NCL-7649 change SCM RepositoryUrl parse functions
